Add dsa signature alg to fips provider
